Very basic, very WIP resource table generator
The beginning of the rabbit-hole has made itself known...

As for `table.json`, it is temporary, but a good example of how the
bootloader will be human-configurable. Most likely, the resource table
will be just one configurable part, and will move into a larger
`config.json` in the root of the repository.

It would be good to be able to include files loaded by UEFI services
at runtime as well (to reduce bootloader executable bloat, and more).
This feature is planned, but the implementation is still fuzzy. I'll
give it some thought as I work on the simple resource table
include-file-in-executable style.