Actually generate `resource_table.[c|h]` from `table.json`
This is a huge step! I'm so happy to finally have a working resource
table being generated based on the JSON configuration file. This means
that users are able to generate their own resource table, in it's own
layout, with custom embedded files and everything. This feels like half
overkill and really cool feature that I really like, haha.