CKernel: Automate CMake
Before there was a whole load of mess that went into generating
`test-build.img` with the kernel installed on it.

Now, the CMake build system inherits the Bootloader CMake system, so
targets like `runimg_qemu` can be invoked, as well as others
extended (for example: `image_raw` is extended to install the
generated kernel executable  into the root directory of the boot
media).

This finally feels like a good system; like it could actually be
imported and used in other projects, not just a proof-of-concept anymore.

Now, the bootloader gets built automatically, but only if it needs to be.