Update top-level CMake
Lots of good stuff in this update.

I spent yesterday solving this CMake dependency stuff (having to
extend targets found in subdirectories is borderline impossible with
how CMake is currently). There is definitely a hole in my door that
wasn't there before, and it suspiciously looks like the outline of my
forehead. Anger + Helplessness = Violence.

All in all, I'm happy with how it turned out.