RADII: Support for `Meson` build system
<https://mesonbuild.com/> is a modern build system
focused on simplicity. One of it's selling points is that
it *isn't* turing complete, and it won't allow itself to
ever be. Custom targets may not run more than a single
shell command. While these are limiting factors, the
overall focus on simplicity does lend itself to a pleasant
end user experience. I even was able to hack on it and get
my PR merged to get this to work at all in a freestanding
UEFI environment.

They are very friendly over there, and I think this build system
can go far, given it's great start on a good community, being fast
(it uses ccache automatically, as well as `mold` linker, etc.).

I haven't implemented the CKernel implementation yet, so I'll do
that now.