CKernel: Remove Meson support
Maybe in the future it would be feasible, but right now it is just too
much work to try and get this working when CMake is just fine and will
continue to be fine.

The real issue lies in that meson subprojects, one, can not have
different compilers, and two, must be in a specific, hard-coded folder
called `subprojects`, and each subdirectory must contain
`meson.build`, and... honestly it's all just too much.

CMake is clean, the industry standard. I would love to give an
underdog like Meson a chance, but I gave them a big change here and it
just doesn't seem to fit.