Meson: Remove entirely
The only reason I didn't revert commits is so that these files are
preserved if I ever want to give Meson another chance. I feel like the
core concept meshes well with what RADII is and has the potential to
be, but the downsides of focusing on simplicity (form) over function
is something that, ultimately, I can't get past. I think if the
inheritance system was better, it was easier to specify cross files
within a meson.build, targets could have multiple commands, and
run-time vs config-time targets/routines with pre/post build options,
etc, then maybe it would be worth it over CMake, but without one or
some of these significant improvements, it's more of a hassle than
anything to support this build system. And I don't see it as worth
it to try and modify/hack on Meson just to get it to work with RADII.